+++ src/jsMain/kotlin/at/posselt/pfrpg2e/Main.kt
@@ -28,6 +28,7 @@ import at.posselt.pfrpg2e.macros.awardXPMacro
 import at.posselt.pfrpg2e.macros.chooseParty
 import at.posselt.pfrpg2e.macros.combatTrackMacro
 import at.posselt.pfrpg2e.macros.createFoodMacro
+import at.posselt.pfrpg2e.macros.createTeleporterPair
 import at.posselt.pfrpg2e.macros.editRealmTileMacro
 import at.posselt.pfrpg2e.macros.editStructureMacro
 import at.posselt.pfrpg2e.macros.resetHeroPointsMacro
@@ -188,7 +189,8 @@ fun main() {
                 editStructureMacro = { actor -> buildPromise { editStructureMacro(actor) } },
                 subsistMacro = { actor -> buildPromise { subsistMacro(game, actor) } },
                 createFoodMacro = { buildPromise { createFoodMacro(game, actionDispatcher) } },
-                showAllNpcHpBars = { buildPromise { game.showAllNpcHpBars() }},
+                showAllNpcHpBarsMacro = { buildPromise { game.showAllNpcHpBars() }},
+                createTeleporterPairMacro = { buildPromise { game.createTeleporterPair() }},
             ),
         )