+++ src/jsMain/kotlin/at/posselt/pfrpg2e/app/forms/ColorInput.kt
@@ -1,0 +1,36 @@
+package at.posselt.pfrpg2e.app.forms
+
+data class ColorInput(
+    override val label: String,
+    override val name: String,
+    val value: String?,
+    val required: Boolean = true,
+    override val help: String? = null,
+    override val hideLabel: Boolean = false,
+    val elementClasses: List<String> = emptyList(),
+    val labelClasses: List<String> = emptyList(),
+    val readonly: Boolean = false,
+    val stacked: Boolean = true,
+    val escapeLabel: Boolean = true,
+    val labelElement: String = "label",
+) : IntoFormElementContext {
+    override fun toContext() = FormElementContext(
+        isFormElement = true,
+        label = label,
+        name = name,
+        help = help,
+        value = value ?: "",
+        required = required,
+        type = "color",
+        disabled = false,
+        readonly = readonly,
+        options = emptyArray(),
+        hideLabel = hideLabel,
+        stacked = stacked,
+        elementClasses = elementClasses.joinToString(" "),
+        labelClasses = labelClasses.joinToString(" "),
+        escapeLabel = escapeLabel,
+        labelElement = labelElement,
+    )
+}
+