+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.1.0] - 2025-04-25
+
 ### Added
 
-* Added a macro that changes all NPCs' Health Bars on all scenes to always display
+* Added a macro that changes all NPCs' Health Bars on all scenes to a configured value
 * Added a macro that creates 2 regions used for teleporting tokens between each other
 
 ## [5.0.0] - 2025-04-21